+++ system-tests/lib/cre/don/config/definitions.go
@@ -176,7 +176,7 @@ func WorkerWorkflowRegistry(workflowRegistryAddr common.Address, homeChainID uin
 	)
 }
 
-func WorkerGateway(nodeAddress common.Address, homeChainID uint64, donID uint32, gatewayConnectorData cre.GatewayConnectorOutput) string {
+func WorkerGateway(nodeAddress common.Address, homeChainID uint64, donID uint64, gatewayConnectorData cre.GatewayConnectorOutput) string {
 	gatewayURL := fmt.Sprintf("ws://%s:%d%s", gatewayConnectorData.Outgoing.Host, gatewayConnectorData.Outgoing.Port, gatewayConnectorData.Outgoing.Path)
 
 	return fmt.Sprintf(`
@@ -189,7 +189,7 @@ func WorkerGateway(nodeAddress common.Address, homeChainID uint64, donID uint32,
 	Id = "por_gateway"
 	URL = "%s"
 `,
-		strconv.FormatUint(uint64(donID), 10),
+		strconv.FormatUint(donID, 10),
 		homeChainID,
 		nodeAddress,
 		gatewayURL,
